move sw_double_and_add_affine and sw_double_and_add_projective to consumers
