+++ lib/crypto/src/curve/short_weierstrass/mod.rs
@@ -7,10 +7,7 @@ mod group;
 pub use group::*;
 
 use crate::{
-    curve::{
-        scalar_mul::{sw_double_and_add_affine, sw_double_and_add_projective},
-        AffineRepr,
-    },
+    curve::AffineRepr,
     field::{group::AdditiveGroup, prime::PrimeField},
 };
 
@@ -106,4 +103,38 @@ pub trait SWCurveConfig: super::CurveConfig {
             .ok_or(bases.len().min(scalars.len()))
     }
     */
+}
+
+/// Standard double-and-add method for multiplication by a scalar.
+#[inline(always)]
+pub fn sw_double_and_add_affine<P: SWCurveConfig>(
+    base: &Affine<P>,
+    scalar: impl AsRef<[u64]>,
+) -> Projective<P> {
+    let mut res = Projective::zero();
+    for b in ark_ff::BitIteratorBE::without_leading_zeros(scalar) {
+        res.double_in_place();
+        if b {
+            res += base
+        }
+    }
+
+    res
+}
+
+/// Standard double-and-add method for multiplication by a scalar.
+#[inline(always)]
+pub fn sw_double_and_add_projective<P: SWCurveConfig>(
+    base: &Projective<P>,
+    scalar: impl AsRef<[u64]>,
+) -> Projective<P> {
+    let mut res = Projective::zero();
+    for b in ark_ff::BitIteratorBE::without_leading_zeros(scalar) {
+        res.double_in_place();
+        if b {
+            res += base
+        }
+    }
+
+    res
 }
